+++ src/class/OutputFormatters/AbstractOutputFormatter.php
@@ -20,6 +20,8 @@
         /**
          * Constructor
          * @author Art <a.molcanovas@gmail.com>
+         * @uses   OutputFormatterStyle::__construct()
+         * @uses   OutputFormatterStyle::setBackground()
          */
         function __construct() {
             $this->formatter = new OutputFormatterStyle();
@@ -39,6 +41,7 @@
          * Static constructor
          * @author Art <a.molcanovas@gmail.com>
          * @return Error|Info|Warning
+         * @uses   AbstractOutputFormatter::__construct()
          */
         static function construct() {
             $class = get_called_class();
@@ -55,6 +58,7 @@
          * @throws InvalidArgumentException When the option name isn't defined
          *
          * @return AbstractOutputFormatter
+         * @uses   OutputFormatterStyle::setOption()
          */
         function setOption($option) {
             $this->formatter->setOption($option);